refactor(compose-app): move CharacterListState into presentation package
